requirements.txt
python-gitlab
python-redmine
